Prospectus_intelligent_review_system_open_source/backend/requirements.txt
pdfplumber==0.11.5
pydantic==2.11.7
python-dateutil==2.9.0.post0
openai>=1.40.0
pandas>=2.2.2
akshare>=1.14.0
python-dotenv>=1.0.1
flask>=3.0.0
flask-cors>=4.0.1
redis>=5.0.7
